Rename avl_set->avl_tree.
