+++ crates/balanced-tree/avl_tree/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "avl-set"
+name = "avl_tree"
 version = "0.1.0"
 edition = "2018"
